requirements.txt
appdirs==1.4.4
attrdict==2.0.1
attrs==22.1.0
base58==2.1.1
blinker==1.4
brotlipy==0.7.0
cattrs==22.1.0
click==8.1.3
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.4.0
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.0.0rc9
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-MySQL==1.5.2
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
hexbytes==0.2.2
install==1.3.5
ipfshttpclient==0.7.0
Jinja2==2.10.3
jsonschema==3.2.0
lru-dict==1.1.8
MarkupSafe==1.1.1
multiaddr==0.0.9
mypy-extensions==0.4.3
mysqlclient==2.1.1
netaddr==0.8.0
omeka-s-tools==0.3.0
packaging==21.3
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
pywin32==304
requests-cache==0.9.6
rlp==1.2.0
six==1.16.0
SQLAlchemy==1.3.24
toolz==0.12.0
typing-extensions==3.10.0.2
url-normalize==1.4.3
varint==1.0.2
web3==5.4.0
websockets==8.1
Werkzeug==0.16.0
wincertstore==0.2
